Use gradle for automatic provision
This allows me to pin down specific versions for Jenkins and all of its
plugins.  This is a lot more reliable since Jenkins rapid development can
easily break the bootstrapping process.

bootstrap.groovy is now skipped because it is unnecessary.  Gradle
handles all of the tasks bootstrap.groovy did except more reliably.  The
only thing gradle doesn't do is update the plugins to the latest version
because this is less reliable.

All plugins are pinned as well.